requirements.txt
google-adk
google-genai
streamlit
reportlab
pyyaml
sqlalchemy
